Solve the issue of the ServiceBean duplication export
