+++ dubbo-config/dubbo-config-api/src/test/java/org/apache/dubbo/config/ServiceConfigTest.java
@@ -136,42 +136,52 @@ public class ServiceConfigTest {
     @Test
     public void testExport() throws Exception {
         service.export();
-
-        assertThat(service.getExportedUrls(), hasSize(1));
-        URL url = service.toUrl();
-        assertThat(url.getProtocol(), equalTo("mockprotocol2"));
-        assertThat(url.getPath(), equalTo(DemoService.class.getName()));
-        assertThat(url.getParameters(), hasEntry(ANYHOST_KEY, "true"));
-        assertThat(url.getParameters(), hasEntry(APPLICATION_KEY, "app"));
-        assertThat(url.getParameters(), hasKey(BIND_IP_KEY));
-        assertThat(url.getParameters(), hasKey(BIND_PORT_KEY));
-        assertThat(url.getParameters(), hasEntry(EXPORT_KEY, "true"));
-        assertThat(url.getParameters(), hasEntry("echo.0.callback", "false"));
-        assertThat(url.getParameters(), hasEntry(GENERIC_KEY, "false"));
-        assertThat(url.getParameters(), hasEntry(INTERFACE_KEY, DemoService.class.getName()));
-        assertThat(url.getParameters(), hasKey(METHODS_KEY));
-        assertThat(url.getParameters().get(METHODS_KEY), containsString("echo"));
-        assertThat(url.getParameters(), hasEntry(SIDE_KEY, PROVIDER));
-        Mockito.verify(protocolDelegate).export(Mockito.any(Invoker.class));
+        try {
+            assertThat(service.getExportedUrls(), hasSize(1));
+            URL url = service.toUrl();
+            assertThat(url.getProtocol(), equalTo("mockprotocol2"));
+            assertThat(url.getPath(), equalTo(DemoService.class.getName()));
+            assertThat(url.getParameters(), hasEntry(ANYHOST_KEY, "true"));
+            assertThat(url.getParameters(), hasEntry(APPLICATION_KEY, "app"));
+            assertThat(url.getParameters(), hasKey(BIND_IP_KEY));
+            assertThat(url.getParameters(), hasKey(BIND_PORT_KEY));
+            assertThat(url.getParameters(), hasEntry(EXPORT_KEY, "true"));
+            assertThat(url.getParameters(), hasEntry("echo.0.callback", "false"));
+            assertThat(url.getParameters(), hasEntry(GENERIC_KEY, "false"));
+            assertThat(url.getParameters(), hasEntry(INTERFACE_KEY, DemoService.class.getName()));
+            assertThat(url.getParameters(), hasKey(METHODS_KEY));
+            assertThat(url.getParameters().get(METHODS_KEY), containsString("echo"));
+            assertThat(url.getParameters(), hasEntry(SIDE_KEY, PROVIDER));
+            Mockito.verify(protocolDelegate).export(Mockito.any(Invoker.class));
+        }finally {
+            service.unexport();
+        }
     }
 
     @Test
     public void testProxy() throws Exception {
         service2.export();
-
-        assertThat(service2.getExportedUrls(), hasSize(1));
-        assertEquals(2, TestProxyFactory.count); // local injvm and registry protocol, so expected is 2
-        TestProxyFactory.count = 0;
+        try {
+            assertThat(service2.getExportedUrls(), hasSize(1));
+            assertEquals(2, TestProxyFactory.count); // local injvm and registry protocol, so expected is 2
+            TestProxyFactory.count = 0;
+        }finally {
+            service2.unexport();
+        }
     }
 
 
     @Test
     public void testDelayExport() throws Exception {
         delayService.export();
-        assertTrue(delayService.getExportedUrls().isEmpty());
-        //add 300ms to ensure that the delayService has been exported
-        TimeUnit.MILLISECONDS.sleep(delayService.getDelay() + 300);
-        assertThat(delayService.getExportedUrls(), hasSize(1));
+        try {
+            assertTrue(delayService.getExportedUrls().isEmpty());
+            //add 300ms to ensure that the delayService has been exported
+            TimeUnit.MILLISECONDS.sleep(delayService.getDelay() + 300);
+            assertThat(delayService.getExportedUrls(), hasSize(1));
+        }finally {
+            delayService.unexport();
+        }
     }
 
     @Test
@@ -260,22 +270,11 @@ public class ServiceConfigTest {
     @Test
     public void testApplicationInUrl() {
         service.export();
-        Assertions.assertNotNull(service.toUrl().getParameter(APPLICATION_KEY));
-        Assertions.assertEquals("app", service.toUrl().getParameter(APPLICATION_KEY));
-    }
-
-
-    @Test
-    public void testDuplicateServiceBean() {
         try {
-            ClassPathXmlApplicationContext ctx = new ClassPathXmlApplicationContext(ServiceConfigTest.class.getPackage().getName().replace('.', '/') + "/demo-provider-duplicate-service-bean.xml");
-            ctx.start();
-            ctx.stop();
-            ctx.close();
-        } catch (IllegalArgumentException e) {
-            Assertions.assertTrue(e.getMessage().contains("The Duplicated BeanDefinition of ServiceBean"));
-            return;
+            Assertions.assertNotNull(service.toUrl().getParameter(APPLICATION_KEY));
+            Assertions.assertEquals("app", service.toUrl().getParameter(APPLICATION_KEY));
+        }finally {
+            service.unexport();
         }
-        Assertions.fail();
     }
 }
