+++ dubbo-config/dubbo-config-spring/src/main/java/org/apache/dubbo/config/spring/ServiceBean.java
@@ -31,6 +31,9 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.context.ApplicationContextAware;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.context.ApplicationEventPublisherAware;
+import org.springframework.context.ApplicationListener;
+import org.springframework.context.event.ContextClosedEvent;
+import org.springframework.context.support.AbstractApplicationContext;
 
 /**
  * ServiceFactoryBean
@@ -38,7 +41,7 @@ import org.springframework.context.ApplicationEventPublisherAware;
  * @export
  */
 public class ServiceBean<T> extends ServiceConfig<T> implements InitializingBean, DisposableBean,
-        ApplicationContextAware, BeanNameAware, ApplicationEventPublisherAware {
+        ApplicationContextAware, BeanNameAware, ApplicationEventPublisherAware, ApplicationListener<ContextClosedEvent> {
 
 
     private static final long serialVersionUID = 213195494150089726L;
@@ -141,5 +144,19 @@ public class ServiceBean<T> extends ServiceConfig<T> implements InitializingBean
     @Override
     public void setApplicationEventPublisher(ApplicationEventPublisher applicationEventPublisher) {
         this.applicationEventPublisher = applicationEventPublisher;
+    }
+
+    @Override
+    public void onApplicationEvent(ContextClosedEvent contextClosedEvent) {
+        super.unexport();
+    }
+
+    @Override
+    protected void delayExport() {
+        if (applicationContext instanceof AbstractApplicationContext) {
+            if (((AbstractApplicationContext) applicationContext).isActive()) {
+                super.delayExport();
+            }
+        }
     }
 }
