+++ dubbo-config/dubbo-config-spring/src/test/resources/META-INF/spring/multiple-services-with-methods.xml
@@ -32,12 +32,12 @@
 
     <!-- service configuration -->
     <dubbo:service interface="org.apache.dubbo.config.spring.api.DemoService"
-                   class="org.apache.dubbo.config.spring.impl.DemoServiceImpl">
+                   class="org.apache.dubbo.config.spring.impl.DemoServiceImpl" group="group1">
         <dubbo:method name="sayName" timeout="500" />
     </dubbo:service>
 
     <dubbo:service interface="org.apache.dubbo.config.spring.api.DemoService"
-                   class="org.apache.dubbo.config.spring.impl.DemoServiceImpl_LongWaiting">
+                   class="org.apache.dubbo.config.spring.impl.DemoServiceImpl_LongWaiting" group="group2">
         <dubbo:method name="sayName" timeout="1000" />
     </dubbo:service>
