+++ dubbo-config/dubbo-config-spring/src/test/resources/org/apache/dubbo/config/spring/demo-provider-duplicate-service-bean.xml
@@ -27,12 +27,12 @@
 
     <dubbo:protocol name="dubbo" port="20887"/>
     <!-- service configuration -->
-    <dubbo:service interface="org.apache.dubbo.config.api.DemoService"
-                   class="org.apache.dubbo.config.provider.impl.DemoServiceImpl">
+    <dubbo:service interface="org.apache.dubbo.config.spring.api.DemoService"
+                   class="org.apache.dubbo.config.spring.impl.DemoServiceImpl" group = "ready_for_duplicate_service_bean">
     </dubbo:service>
 
-    <dubbo:service interface="org.apache.dubbo.config.api.DemoService"
-                   class="org.apache.dubbo.config.provider.impl.DemoServiceImpl">
+    <dubbo:service interface="org.apache.dubbo.config.spring.api.DemoService"
+                   class="org.apache.dubbo.config.spring.impl.DemoServiceImpl" group = "ready_for_duplicate_service_bean">
     </dubbo:service>
 
 </beans>
